refactor: the baseUrl of tsconfig
