+++ src/utils/index.ts
@@ -1,6 +1,6 @@
-import extractFlags from "./extractFlags";
-import getRepoName from "./getRepoName";
-import getGitAuthor from "./getGitAuthor";
-import configurePackageJson from "./configurepackagejson";
+import extractFlags from "@/src/utils/extractFlags";
+import getRepoName from "@/src/utils/getRepoName";
+import getGitAuthor from "@/src/utils/getGitAuthor";
+import configurePackageJson from "@/src/utils/configurepackagejson";
 
 export { extractFlags, getRepoName, getGitAuthor, configurePackageJson };
