+++ src/utils/logger.ts
@@ -1,0 +1,19 @@
+import chalk from "chalk";
+
+export const logger = {
+  error(...args: unknown[]) {
+    console.log(chalk.red(...args));
+  },
+  warn(...args: unknown[]) {
+    console.log(chalk.yellow(...args));
+  },
+  info(...args: unknown[]) {
+    console.log(chalk.cyan(...args));
+  },
+  success(...args: unknown[]) {
+    console.log(chalk.green(...args));
+  },
+  break() {
+    console.log("");
+  },
+};
